Create client and tutor models.
First step in implementing appointments.  The Tutor and Client models
allow for a given User to be both a Tutor and a Client to some other
Tutor.

Ex:  I may tutor math, and be tutored in chemistry.

The other implementation option is an appointment model with multiple
foreign keys to User.  However, we have additional metadata to store
with tutors (i.e., subjects tutored), which supports using an extra
table.

(Also, multiple FKs to the same table are ugly in Rails.)