Tutor schedule for day.
Tutor::TutorDaySchedule is intended to simplify the view logic listing a
tutor's full day schedule.  It is an iterator that yields hashes, each
corresponding to a single cell in the schedule table.  These hashes detail
the start time, length in cells, and type of block, so that the view code
doesn't have to care.