Added capybara-webkit.
But don't actually activate it.  Using Webkit causes our database cleaner
after hook to crash (database is locked).  Most likely because SQLite3 is
horrible at handling multiple connections, and Webkit runs fast enough to
uncover a race condition.

So for now, going to live with a slower test suite.  Upgrading test and
development databases to Postgres *might* solve the issue, but that's
a non-trivial amount of work.